# contracts/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Condux Control Plane API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8099/"
    }
  ],
  "paths": {
...
    "/api/github/webhook": {
      "post": {
        "tags": [
          "GitHub"
        ],
        "operationId": "githubWebhook",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/github/setup": {
      "get": {
        "tags": [
          "GitHub"
        ],
        "operationId": "githubSetup",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/orgs/{orgId}/github": {
      "get": {
        "tags": [
          "GitHub"
        ],
        "operationId": "listGithubInstallations",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GithubInstallationResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "GitHub"
        ],
        "operationId": "disconnectGithub",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/orgs/{orgId}/github/repositories": {
      "get": {
        "tags": [
          "GitHub"
        ],
        "operationId": "listGithubRepositories",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubRepositoriesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/orgs/{orgId}/github/branches": {
      "get": {
        "tags": [
          "GitHub"
        ],
        "operationId": "listGithubBranches",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubBranchesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GithubBranchesResponse": {
        "required": [
          "branches"
        ],
        "type": "object",
        "properties": {
          "branches": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "GithubInstallationResponse": {
        "required": [
          "installationId",
          "accountLogin",
          "createdAt",
          "manageUrl"
        ],
        "type": "object",
        "properties": {
          "installationId": {
            "type": "integer",
            "format": "int64"
          },
          "accountLogin": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "manageUrl": {
            "type": "string"
          }
        }
      },
      "GithubRepositoriesResponse": {
        "required": [
          "repositories"
        ],
        "type": "object",
        "properties": {
          "repositories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}